pyproject.toml
[tool.poetry]
name = "courierсrm"
version = "0.1.0"
description = ""
authors = ["RomanL91 <dev_roman91@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.7"
fastapi = "^0.115.11"
uvicorn = "^0.34.0"
requests = "^2.32.3"
pika = "^1.3.2"
aiogram = "^3.18.0"
aio-pika = "^9.5.5"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
